HMAN-511, HMAN-512, HMAN-513 sonar issues
