+++ src/test/java/uk/gov/hmcts/reform/hmc/service/LinkHearingGroupServiceTest.java
@@ -120,13 +120,15 @@ class LinkHearingGroupServiceTest {
     LinkedHearingDetailsAuditRepository linkedHearingDetailsAuditRepository;
 
     @Mock
-    private LinkedGroupDetailsAuditRepository linkedGroupDetailsAuditRepository;
+    LinkedGroupDetailsAuditRepository linkedGroupDetailsAuditRepository;
 
     @Mock
-    private LinkedGroupDetailsAuditMapper linkedGroupDetailsAuditMapper;
+    LinkedGroupDetailsAuditMapper linkedGroupDetailsAuditMapper;
 
     @Mock
-    private LinkedHearingDetailsAuditMapper linkedHearingDetailsAuditMapper;
+    LinkedHearingDetailsAuditMapper linkedHearingDetailsAuditMapper;
+
+    FutureHearingsLinkedHearingGroupService futureHearingsLinkedHearingGroupService;
 
     HearingIdValidator hearingIdValidator;
 
@@ -142,21 +144,32 @@ class LinkHearingGroupServiceTest {
     public void setUp() {
         MockitoAnnotations.openMocks(this);
         hearingIdValidator = new HearingIdValidator(hearingRepository, actualHearingRepository,
-                actualHearingDayRepository);
+                                                    actualHearingDayRepository
+        );
         linkedHearingValidator = new LinkedHearingValidator(hearingIdValidator, hearingRepository,
-                linkedGroupDetailsRepository, linkedHearingDetailsRepository);
+                                                            linkedGroupDetailsRepository, linkedHearingDetailsRepository
+        );
+        futureHearingsLinkedHearingGroupService =
+            new FutureHearingsLinkedHearingGroupService(
+                hearingRepository,
+                linkedGroupDetailsRepository,
+                linkedHearingValidator,
+                linkedHearingDetailsAuditRepository,
+                linkedGroupDetailsAuditRepository,
+                linkedGroupDetailsAuditMapper,
+                linkedHearingDetailsAuditMapper,
+                accessControlService
+            );
         service =
-            new LinkedHearingGroupServiceImpl(hearingRepository,
-                                              linkedGroupDetailsRepository,
-                                              linkedHearingValidator,
-                                              linkedHearingDetailsAuditRepository,
-                                              linkedGroupDetailsAuditRepository,
-                                              linkedGroupDetailsAuditMapper,
-                                              linkedHearingDetailsAuditMapper,
-                                              futureHearingRepository,
-                                              objectMapper,
-                                              accessControlService
-                                              );
+            new LinkedHearingGroupServiceImpl(
+                hearingRepository,
+                linkedGroupDetailsRepository,
+                linkedHearingValidator,
+                futureHearingRepository,
+                objectMapper,
+                accessControlService,
+                futureHearingsLinkedHearingGroupService
+            );
     }
 
     @Nested
@@ -209,7 +222,6 @@ class LinkHearingGroupServiceTest {
             assertEquals("001 Insufficient requestIds", exception.getMessage());
         }
 
-
         @Test
         void shouldFailWithHearingRequestIsLinkedIsFalse() {
             GroupDetails groupDetails = generateGroupDetails("comment", "name",
@@ -244,7 +256,6 @@ class LinkHearingGroupServiceTest {
             assertEquals("002 hearing request isLinked is False", exception.getMessage());
         }
 
-
         @Test
         void shouldFailWithHearingRequestAlreadyInGroup() {
             LinkedGroupDetails groupDetails1 = generateLinkGroupDetails(
@@ -259,15 +270,15 @@ class LinkHearingGroupServiceTest {
                 1L
             );
             LinkedGroupDetails groupDetailsAlternate = generateLinkGroupDetails(
-                    202L,
-                    "requestId2",
-                    "request name2",
-                    "Same Slot",
-                    "status",
-                    "reason",
-                    "comments",
-                    LocalDateTime.of(2022, 03, 02, 10, 11),
-                    1L
+                202L,
+                "requestId2",
+                "request name2",
+                "Same Slot",
+                "status",
+                "reason",
+                "comments",
+                LocalDateTime.of(2022, 03, 02, 10, 11),
+                1L
             );
 
             HearingEntity hearingEntity = generateHearingEntity(
@@ -290,16 +301,16 @@ class LinkHearingGroupServiceTest {
             when(hearingRepository.findById(any())).thenReturn(Optional.of(hearingEntity));
 
             GroupDetails groupDetails = generateGroupDetails("comment", "name",
-                    LinkType.ORDERED.label, "reason"
+                                                             LinkType.ORDERED.label, "reason"
             );
 
             LinkHearingDetails hearingDetails2 = generateHearingDetails("2000000002", 2);
             HearingLinkGroupRequest hearingLinkGroupRequest = generateHearingLink(
-                    groupDetails,
-                    Arrays.asList(
-                            hearingDetails1,
-                            hearingDetails2
-                    )
+                groupDetails,
+                Arrays.asList(
+                    hearingDetails1,
+                    hearingDetails2
+                )
             );
 
             Exception exception = assertThrows(BadRequestException.class, () -> {
@@ -307,7 +318,6 @@ class LinkHearingGroupServiceTest {
             });
             assertEquals("003 hearing request already in a group", exception.getMessage());
         }
-
 
         @Test
         void shouldFailWithInvalidState() {
@@ -549,7 +559,6 @@ class LinkHearingGroupServiceTest {
             verify(hearingRepository, times(2)).save(any());
             verify(linkedGroupDetailsRepository, times(2)).save(any());
         }
-
 
         @Test
         void shouldPassWhenHearingOrderIsSameSlotWithOrderZero() {
@@ -1248,8 +1257,10 @@ class LinkHearingGroupServiceTest {
             errorDetails.setErrorCode(400);
             BadFutureHearingRequestException badFutureHearingRequestException = new BadFutureHearingRequestException(
                 INVALID_REQUEST);
-            doThrow(badFutureHearingRequestException).when(futureHearingRepository).updateLinkedHearingGroup(any(),
-                                                                                                             any());
+            doThrow(badFutureHearingRequestException).when(futureHearingRepository).updateLinkedHearingGroup(
+                any(),
+                any()
+            );
 
             Exception exception = assertThrows(BadRequestException.class, () -> {
                 service.updateLinkHearing("1", hearingLinkGroupRequest);
